backend/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv

# AI/ML dependencies
groq
xgboost
scikit-learn
shap
joblib
pandas
numpy

# OCR and image processing
rapidocr-onnxruntime
opencv-python
Pillow
pdf2image
rapidfuzz
python-multipart

# Blockchain and document generation
reportlab
qrcode[pil]
cryptography

# Additional utilities
pydantic>=2.0.0
httpx
